API to get a single IP's data
This patch implements the method to get the data for a single IP
address.

The patch keeps the old API paths to avoid changing more than necessary.
This does lead to the necessity of using URL parameters for specifying
the requested IP address. Mid-term, it probably makes more sense to turn
the IP address into a path parameter for all endpoints operating on a
single IP address.