Fix tag conflicts
This patch fixes a bug in handling several tags where the API would
sometimes complain about already existing tags and sometimes wouldn't.

This will work just fine:
```
❯ deterrers-cli add --admin lkiesow --admin virtUOS 131.173.20.15
```

This will result in a 409 Conflict:
```
❯ deterrers-cli add --admin virtUOS --admin lkiesow 131.173.20.15
```

Worse, this will not only result in a conflict but also leave a half
added entity around in DETERRERS.

If someone wants a tag to be added it should be fine to accept the
request and return success if the tag already exists.