Update List of Admins
This patch allows users to update the list of admins for an IP address
along with the other information.

Not only allows this to add or remove an admin without going through the
whole process of unregistering the host, removing the host and then
adding the host again, it also means that add and update now actually
take the same data as input.