Ensure users can tell that a host was blocked
Registering a number of hosts, it is impossible to tell if any of the
hosts was blocked without carefully coming though all of them since the
subject is the same regardless of the status.

This patch makes DETERRERS send out a different title when a
vulnerability was detected in the initial scan and the host was blocked.